Add files via upload
The original notebook is running without errors and is producing the expected results. The text is clear and relatively free of jargon. I corrected a few spelling and grammar errors.
I added reference to Lorenz 96 paper.
In class GCM there are 2 commented lines of code which replicate what is coded in other lines that are kept. I removed those commented lines for clarity.
I added axis-labels and legends for the figures.