L96 description and references
I added a description of the L96 model in the Jupyter notebook 'presentation-model-setup.ipynb' as it is used now as the first chapter of the Jupyter Book. I also included various references to numerical methods developed for the L96 system. 
Please let me know if there is any suggestion to improve the notebook.